basic alias expansion now implemented
each word passed to sh_eval now gets checked to see if it matches an alias, if it does it gets expanded. There is no support for positional paramters to aliases yet

both sh and csh style alias syntax is allowed, e.g. "alias la=ls -a" or "alias la ls -a". Note that the shell has no concept of "quotes" yet