Code now respects $IFS environment variable correctly.
Before there were bits and pieces of code that used whatever felt good at the moment for field spliting command lines; other parts respected $IFS in a bug prone manor.

The new ifsplit LIST operator will split it's arguments IAW with the users $IFS (or a default if not set) and return the resulting list.

The default value for $ENV{IFS} in init_sh() will likely be removed in the near future.