do_completion subroutine added
a basic function for tab completion; does filename, builtin command, and
alias completion. Still more work to do, but it's a start.

To use it with the Term::ReadLine::Zoid library, set

	PERL_RL="Zoid completion_function=main::do_completion"

dealing with the other (crappy) implementations of ReadLine is not done
yet. Puggable hooks and several completions are not done yet.