initial implementation of the function and macro built-ins
'function name() definition' creates an alias named 'name' with evaluation
'definition'

'macro name( definition )' creates an alias named 'name' with evaluation
'definition'

Eventually macro will probably define the alias wrapped in a (grouping),
once (command groupings) are implemented. function will allow {
groupings } when that is also implemented; but will not make the
implicit.

Currently variables and quotes are pre-expanded by the line processing
code, before function/macro receive the text to store in a macro. While
this is the desired behavour for the alias built in, it is not desired
for other methods of defining a command macro! It is however possible to
escape variables and things, e.g.$ macro e( echo hi '$USER' ); so that
the expansion occurs when 'e' is executed, as opposed to when 'e' is
recorded. This will hopefully be fixed in the future.

The manual page is in continual development, and will reflect the
desired "end state" for macro behavour.