tpsh < script and echo 'cmds' | tpsh now work correctly
after argument processing, tpsh now checks if *STDIN is connected to a
terminal and sets 'interactive' accordingly; also 'stdin' is set IAW
with the contents of @ARGV after argument processing. Left over arguments are
assumed to be files, no checking is done. The reason for this, I do not
wish to stat each argument and *then* also have to process what to do
with it; thus it is the users problem if they call us 'tpsh rm file'
instead of 'tpsh -c "rm file"'.

the xtrace option (-x) is now implemented imperfectly but suitably.