do_getop() made generic, `set [options]` now works.
The do_getopt() function now takes an array ref and hash to pass onto
Getopt::Long::GetOptionsFromArray, and twists it in place.

set_bin() now calls do_getopt() with it's argument vector and a global
hash; startup code now calls do_getopt() with \@ARGV and that same hash.
Because of that, `set -x`, `set -o xtrace`, and such work but `set +o
xtrace` and such can't be used to turn the option off yet lol. (+o atm
is just an alias for -o.)