improved macro expansion
	alias ls="ls -FGH"
	alias lsa='ls $HOME'

lsa now expands to ls -FGH $HOME instead of ls $HOME which is then
executed as `ls -FGH /user/home/Terry` in my case.

Recursive macros are now handled in the following way:

	$ alias x y
	$ alias y z
	$ alias z x
	$ echo x
	z
	$ echo y
	x
	$ echo z
	y
	$

At the moment, after a -> b -> c, it stops; so a -> b -> c -> d -> e ->
f -> a, resolves to 'c' instead of 'f'.