fc built-in command implemented
Doing this has given me new respect for perl coderefs (if not
syntactically), as well as fc(1) and line editors (ed(1), ex(1)). The
command has been implemented in ~200 lines of code. The only real thing
that needs doing is to make it able to accept negative / positive array
indexes into the history; currently do_getopt() doesn't provide a way to
touch Getopt::Long::Configure(). We'll have to change that ;-)

Arguments to fc may be array indexes or perl regular expressions. The
question is, should we drop to a simple substr() driven operation: so we
can skip the overhead of regexp and avoid the need for the eval{ is this a
valid regexp } thing.