improved quote handling
quotes now escape things and can be nested properly; unclosed quotes are
not handled yet, so they just get returned. (Fixing this is on the todo
list). Better delegation between new tpsh_parse and old tpsh_lex.

Code development is ON GOING; the parserlexer branch is for unstable
development of tpsh's parsing and lexing related code.