lexer understands ;, &, and |.
The parser takes care of simple quoting and field splitting.
The lexer takes care of expanding the resulting tokens, and returns the
execution queue.

If the queue basically takes `cmd1; cmd2; ...` and turns it into that
kind of a list of commands to run. If the parser output includes an '&'
it is joined to the previous token: pexec() or resolve_cmd() will have
to learn to understand the async operator (&). If the parser output
includes a '|', it is joined to the end of the previous token: as a sign
to resolve_cmd to join each command with the pipe operator (|) with the
next command in the quenue, until it runs out of piping.

Note: debugging statements remain in this commit.