refactoring of resolve_cmd() begins
$0 is now overrated using FindBin and File::Spec; so things like
err_handler() and death_handler() need to localize and trip $0; which
they now do (obviously). This is done so we can do "$0 -c args" as a
child process on things like `backticks` and $(command substations).

Pipes are currently disabled, since they were intentionally broken in the merge
anyway.

I/O redirection now works though the argument vector syntax rather then
lexically grepping the line: if the -2'th argument is a redirection
operator, then the -1'th argument is the file; and the other arguments
is the command set to run. tpsh now checks for a file descriptor, e.g.
n<, n>, and n>>; but doesn't do anything with it *yet*.