Rewrote init_sh
The init code is a little clearner now, and separates windows/unix
differences to separate subroutines.  Also the UserGetInfo subroutine
from Win32API::Net is no longer required for Win32 support but is
strongly reccomended. This was done because it isn't bundled with
Strawberry perl.

Also note that there is no longer a default $PATH setting and that
LOGNAME/USER/USERNAME handling is more relaxed.