Finish rewrite of fc -l to take regexes as indexes
I would like for something like fc -l regex index or fc -l index regex
to work as intended, but I don't see any way that it can be bullet
proof.

Saying if first is a regex, second must be as well: means you can't make
the second an index.

Saying if second looks like an index, you can't use a purely numeric
pattern!

I will have to think on that a little.